+++ src/app/(post-login)/submission/program/[shortName]/clinical-data/ClinicalEntityDataTable/dataQuery.tsx
@@ -1,0 +1,55 @@
+/*
+ * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ *
+ * This program and the accompanying materials are made available under the terms of
+ * the GNU Affero General Public License v3.0. You should have received a copy of the
+ * GNU Affero General Public License along with this program.
+ *  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
+ * SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
+ * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+import CLINICAL_ENTITY_DATA_QUERY from '@/app/gql/clinical/CLINICAL_ENTITY_DATA_QUERY';
+import { useClinicalQuery } from '@/app/hooks/useApolloQuery';
+import { CompletionStates, clinicalEntityFields, defaultClinicalEntityFilters } from './common';
+
+const validateEntityQueryName = (entityQuery) => {
+	const entities = typeof entityQuery === 'string' ? [entityQuery] : entityQuery;
+	return entities.map((entityName) => clinicalEntityFields.find((entity) => entity === entityName));
+};
+
+export const useGetEntityData = (
+	program: string,
+	entityType: string | string[],
+	page: number,
+	pageSize: number,
+	sort: string,
+	completionState: CompletionStates,
+	donorIds: number[],
+	submitterDonorIds: string[],
+) =>
+	useClinicalQuery(CLINICAL_ENTITY_DATA_QUERY, {
+		errorPolicy: 'all',
+		fetchPolicy: 'cache-and-network',
+		variables: {
+			programShortName: program,
+			filters: {
+				...defaultClinicalEntityFilters,
+				sort,
+				page,
+				pageSize,
+				completionState,
+				donorIds,
+				submitterDonorIds,
+				entityTypes: validateEntityQueryName(entityType),
+			},
+		},
+	});
