+++ src/app/(post-login)/submission/program/[shortName]/clinical-data/ClinicalEntityDataTable/types.ts
@@ -1,0 +1,5 @@
+export type DonorEntry = {
+	row: string;
+	isNew: boolean;
+	[k: string]: string | number | boolean;
+};
